[ACPI] update Linux to ACPICA 20031029 (Bob Moore)
Fixed a problem where a level-triggered GPE with an associated _Lxx
control method was incorrectly cleared twice.

Fixed a problem with the Field support code where an access can occur
beyond the end-of-region if the field is non-aligned but extends to the
very end of the parent region (resulted in an AE_AML_REGION_LIMIT
exception.)

Fixed a problem with ACPI Fixed Events where an RT Clock handler would
not get invoked on an RTC event.  The RTC event bitmasks for the PM1
registers were not being initialized properly.

Implemented support for executing _STA and _INI methods for Processor
objects.  Although this is currently not part of the ACPI specification,
there is existing ASL code that depends on the init-time execution of
these methods.

Implemented and deployed a GetDescriptorName function to decode the
various types of internal descriptors.  Guards against null descriptors
during debug output also.

Implemented and deployed a GetNodeName function to extract the
4-character namespace node name.  This function simplifies the debug and
error output, as well as guarding against null pointers during output.

Implemented and deployed the ACPI_FORMAT_UINT64 helper macro to simplify
the debug and error output of 64-bit integers.  This macro replaces the
HIDWORD and LODWORD macros for dumping these integers.

Updated the implementation of the Stall() operator to only call
AcpiOsStall(), and also return an error if the operand is larger than
255.  This preserves the required behavior of not relinquishing the
processor, as would happen if AcpiOsSleep() was called for "long
stalls".

Constructs of the form "Store(LocalX,LocalX)" where LocalX is not
initialized are now treated as NOOPs.

Cleaned up a handful of warnings during 64-bit generation.

Fixed a reported error where and incorrect GPE number was passed to the
GPE dispatch handler.  This value is only used for error output,
however.  Used this opportunity to clean up and streamline the GPE
dispatch code.

+++ drivers/acpi/parser/psargs.c
@@ -490,7 +490,7 @@ union acpi_parse_object *
 acpi_ps_get_next_field (
 	struct acpi_parse_state         *parser_state)
 {
-	u32                             aml_offset = ACPI_PTR_DIFF (parser_state->aml,
+	u32                             aml_offset = (u32) ACPI_PTR_DIFF (parser_state->aml,
 			 parser_state->aml_start);
 	union acpi_parse_object         *field;
 	u16                             opcode;
@@ -677,7 +677,7 @@ acpi_ps_get_next_arg (
 
 			/* Fill in bytelist data */
 
-			arg->common.value.size = ACPI_PTR_DIFF (parser_state->pkg_end,
+			arg->common.value.size = (u32) ACPI_PTR_DIFF (parser_state->pkg_end,
 					  parser_state->aml);
 			arg->named.data = parser_state->aml;
